+++ src/commands/history.ts
@@ -2,7 +2,9 @@
 
 import * as vscode from 'vscode';
 
-import { DiffDialog } from '../services/common/DiffDialog';
+import * as commands from '../common/commands';
+import * as dialogs from '../common/dialogs';
+
 import { DiffMenu } from '../services/panel/DiffMenu';
 import { DiffPanel } from '../services/panel/DiffPanel';
 import { DiffHistory } from '../services/sidebar/DiffHistory';
@@ -23,42 +25,45 @@ export function activate (context:vscode.ExtensionContext) {
 	
 	vscode.window.registerTreeDataProvider('l13DiffHistory', diffHistoryProvider);
 	
-	context.subscriptions.push(vscode.commands.registerCommand('l13Diff.openComparison', ({ comparison }) => {
+	commands.register(context, {
 		
-		DiffPanel.createOrShow(context, [{ fsPath: comparison.fileA }, { fsPath: comparison.fileB }], true);
+		'l13Diff.openComparison': ({ comparison }) => {
+			
+			DiffPanel.createOrShow(context, [{ fsPath: comparison.fileA }, { fsPath: comparison.fileB }], true);
+			
+		},
 		
-	}));
-	
-	context.subscriptions.push(vscode.commands.registerCommand('l13Diff.openComparisonOnly', ({ comparison }) => {
+		'l13Diff.openComparisonOnly': ({ comparison }) => {
+			
+			DiffPanel.createOrShow(context, [{ fsPath: comparison.fileA }, { fsPath: comparison.fileB }], false);
+			
+		},
 		
-		DiffPanel.createOrShow(context, [{ fsPath: comparison.fileA }, { fsPath: comparison.fileB }], false);
+		'l13Diff.openComparisonAndCompare': ({ comparison }) => {
+			
+			DiffPanel.createOrShow(context, [{ fsPath: comparison.fileA }, { fsPath: comparison.fileB }], true);
+			
+		},
 		
-	}));
-	
-	context.subscriptions.push(vscode.commands.registerCommand('l13Diff.openComparisonAndCompare', ({ comparison }) => {
+		'l13Diff.openComparisonInNewPanel': ({ comparison }) => {
+			
+			DiffPanel.create(context, [{ fsPath: comparison.fileA }, { fsPath: comparison.fileB }], true);
+			
+		},
 		
-		DiffPanel.createOrShow(context, [{ fsPath: comparison.fileA }, { fsPath: comparison.fileB }], true);
+		'l13Diff.removeComparison': ({ comparison }) => DiffHistory.removeComparison(context, comparison),
 		
-	}));
-	
-	context.subscriptions.push(vscode.commands.registerCommand('l13Diff.openComparisonInNewPanel', ({ comparison }) => {
-		
-		DiffPanel.create(context, [{ fsPath: comparison.fileA }, { fsPath: comparison.fileB }], true);
-		
-	}));
-	
-	context.subscriptions.push(vscode.commands.registerCommand('l13Diff.removeComparison', ({ comparison }) => DiffHistory.removeComparison(context, comparison)));
-	
-	context.subscriptions.push(vscode.commands.registerCommand('l13Diff.clearHistory', async () => {
-		
-		const value = await DiffDialog.confirm('Delete the complete history?', 'Delete');
-		
-		if (value) {
-			DiffMenu.clearHistory(context);
-			DiffHistory.clearComparisons(context);
-		}
-		
-	}));
+		'l13Diff.clearHistory': async () => {
+			
+			const value = await dialogs.confirm('Delete the complete history?', 'Delete');
+			
+			if (value) {
+				DiffMenu.clearHistory(context);
+				DiffHistory.clearComparisons(context);
+			}
+			
+		},
+	});
 	
 }