+++ src/services/actions/DiffCopy.ts
@@ -4,16 +4,16 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
 
-import { copyFile, lstatSync, mkdirsSync } from '../@l13/nodes/fse';
+import { copyFile, lstatSync, mkdirsSync } from '../@l13/fse';
 
 import { CopyFileEvent, CopyFilesEvent, CopyFilesJob, Diff, DiffCopyMessage, DiffFile, DiffMultiCopyMessage, MultiCopyEvent } from '../../types';
 
-import { DiffDialog } from '../common/DiffDialog';
-import { DiffSettings } from '../common/DiffSettings';
+import * as dialogs from '../../common/dialogs';
+import * as settings from '../../common/settings';
 
 //	Variables __________________________________________________________________
 
-const BUTTON_COPY_DONT_ASK_AGAIN = 'Copy, don\'t ask again';
+const BUTTON_COPY_DONT_SHOW_AGAIN = 'Copy, don\'t show again';
 
 //	Initialize _________________________________________________________________
 
@@ -91,19 +91,19 @@ export class DiffCopy {
 	
 	public async showCopyFromToDialog (data:DiffCopyMessage, from:'A'|'B', to:'A'|'B') {
 		
-		const confirmCopy = DiffSettings.get('confirmCopy', true);
+		const confirmCopy = settings.get('confirmCopy', true);
 		const length = data.diffs.length;
 		
 		if (!length) return;
 		
 		if (confirmCopy && !data.multi) {
 			const text = `Copy ${length > 1 ? length + ' files' : `"${data.diffs[0].id}"`} to "${(<any>data)['path' + to]}"?`;
-			const value = await DiffDialog.confirm(text, 'Copy', BUTTON_COPY_DONT_ASK_AGAIN);
+			const value = await dialogs.confirm(text, 'Copy', BUTTON_COPY_DONT_SHOW_AGAIN);
 				
 			if (value) {
-				if (value === BUTTON_COPY_DONT_ASK_AGAIN) DiffSettings.update('confirmCopy', false);
+				if (value === BUTTON_COPY_DONT_SHOW_AGAIN) settings.update('confirmCopy', false);
 				this.copyFromTo(data, from, to);
-			} else this._onDidCancel.fire();
+			} else this._onDidCancel.fire(undefined);
 		} else this.copyFromTo(data, from, to);;
 		
 	}
@@ -117,10 +117,10 @@ export class DiffCopy {
 		
 		const folderFrom = from === 'left' ? data.pathA : data.pathB;
 		const text = `Copy ${length > 1 ? length + ' files' : `"${ids[0]}"`} from "${folderFrom}" across all diff panels?`;
-		const value = await DiffDialog.confirm(text, 'Copy');
+		const value = await dialogs.confirm(text, 'Copy');
 		
 		if (value) this._onInitMultiCopy.fire({ data, from });
-		else this._onDidCancel.fire();
+		else this._onDidCancel.fire(undefined);
 		
 	}
 	
@@ -148,7 +148,7 @@ export class DiffCopy {
 				
 				try {
 					await this.copy(fileFrom, dest);
-					diff.status = 'unchanged';
+					if (diff.status !== 'ignored') diff.status = 'unchanged';
 					let fileTo = to === 'A' ? diff.fileA : diff.fileB;
 				
 					if (!fileTo) {
@@ -164,6 +164,7 @@ export class DiffCopy {
 							dirname: fileFrom.dirname,
 							extname: fileFrom.extname,
 							type: fileFrom.type,
+							ignore: fileFrom.ignore,
 						};
 						if (to === 'A') diff.fileA = fileTo;
 						else diff.fileB = fileTo;