+++ src/services/sidebar/DiffFavorites.ts
@@ -2,11 +2,12 @@
 
 import * as vscode from 'vscode';
 
-import { sortCaseInsensitive } from '../../@l13/natvies/arrays';
-import { Favorite, FavoriteGroup, InitialState } from '../../types';
+import { sortCaseInsensitive } from '../../@l13/arrays';
+import { Favorite, FavoriteGroup, FavoriteTreeItems, InitialState } from '../../types';
 
-import { DiffDialog } from '../common/DiffDialog';
-import { DiffSettings } from '../common/DiffSettings';
+import * as dialogs from '../../common/dialogs';
+import * as settings from '../../common/settings';
+
 import { FavoriteGroupTreeItem } from './trees/FavoriteGroupTreeItem';
 import { FavoriteTreeItem } from './trees/FavoriteTreeItem';
 
@@ -16,8 +17,6 @@ const FAVORITES = 'favorites';
 const FAVORITE_GROUPS = 'favoriteGroups';
 
 const BUTTON_DELETE_GROUP_AND_FAVORITES = 'Delete Group and Favorites';
-
-type FavoriteTreeItems = FavoriteTreeItem|FavoriteGroupTreeItem;
 
 //	Initialize _________________________________________________________________
 
@@ -48,7 +47,7 @@ export class DiffFavorites implements vscode.TreeDataProvider<FavoriteTreeItems>
 		this.favorites = this.context.globalState.get(FAVORITES) || [];
 		this.favoriteGroups = this.context.globalState.get(FAVORITE_GROUPS, []);
 		
-		const initialState:InitialState = DiffSettings.get('initialFavoriteGroupState', 'Remember');
+		const initialState:InitialState = settings.get('initialFavoriteGroupState', 'Remember');
 		
 		if (initialState !== 'Remember') {
 			this.favoriteGroups.forEach((favoriteGroup) => favoriteGroup.collapsed = initialState === 'Collapsed');
@@ -96,20 +95,6 @@ export class DiffFavorites implements vscode.TreeDataProvider<FavoriteTreeItems>
 		
 	}
 	
-	public collapseAll () {
-		
-		const favoriteGroups:FavoriteGroup[] = this.context.globalState.get(FAVORITE_GROUPS, []);
-		
-		FavoriteGroupTreeItem.updateStateVersion();
-		
-		favoriteGroups.forEach((favoriteGroup) => favoriteGroup.collapsed = true);
-		
-		this.context.globalState.update(FAVORITE_GROUPS, favoriteGroups);
-		
-		this.refresh();
-		
-	}
-	
 	public static async addFavorite (context:vscode.ExtensionContext, fileA:string, fileB:string) {
 		
 		const label = await vscode.window.showInputBox({
@@ -123,7 +108,7 @@ export class DiffFavorites implements vscode.TreeDataProvider<FavoriteTreeItems>
 		
 		for (const favorite of favorites) {
 			if (favorite.label === label) {
-				if (!await DiffDialog.confirm(`Overwrite favorite "${favorite.label}"?`, 'Ok')) return;
+				if (!await dialogs.confirm(`Overwrite favorite "${favorite.label}"?`, 'Ok')) return;
 				favorite.fileA = fileA;
 				favorite.fileB = fileB;
 				return saveFavorite(context, favorites);;
@@ -168,9 +153,7 @@ export class DiffFavorites implements vscode.TreeDataProvider<FavoriteTreeItems>
 	
 	public static async removeFavorite (context:vscode.ExtensionContext, favorite:Favorite) {
 		
-		const value = await DiffDialog.confirm(`Delete favorite "${favorite.label}"?`, 'Delete');
-		
-		if (value) {
+		if (await dialogs.confirm(`Delete favorite "${favorite.label}"?`, 'Delete')) {
 			const favorites:Favorite[] = context.globalState.get(FAVORITES) || [];
 			
 			for (let i = 0; i < favorites.length; i++) {
@@ -298,7 +281,7 @@ export class DiffFavorites implements vscode.TreeDataProvider<FavoriteTreeItems>
 	
 	public static async removeFavoriteGroup (context:vscode.ExtensionContext, favoriteGroup:FavoriteGroup) {
 		
-		const value = await DiffDialog.confirm(`Delete favorite group "${favoriteGroup.label}"?`, 'Delete', BUTTON_DELETE_GROUP_AND_FAVORITES);
+		const value = await dialogs.confirm(`Delete favorite group "${favoriteGroup.label}"?`, 'Delete', BUTTON_DELETE_GROUP_AND_FAVORITES);
 		
 		if (value) {
 			const favoriteGroups:FavoriteGroup[] = context.globalState.get(FAVORITE_GROUPS, []);
@@ -329,9 +312,7 @@ export class DiffFavorites implements vscode.TreeDataProvider<FavoriteTreeItems>
 	
 	public static async clearFavorites (context:vscode.ExtensionContext) {
 		
-		const value = await DiffDialog.confirm(`Delete all favorites and groups?'`, 'Delete');
-		
-		if (value) {
+		if (await dialogs.confirm(`Delete all favorites and groups?'`, 'Delete')) {
 			context.globalState.update(FAVORITES, []);
 			context.globalState.update(FAVORITE_GROUPS, []);
 			DiffFavorites.currentProvider?.refresh();