+++ src/views/@l13/events/event-dispatcher.class.ts
@@ -1,6 +1,6 @@
 //	Imports ____________________________________________________________________
 
-import { remove } from '../../../@l13/natvies/arrays';
+import { remove } from '../../../@l13/arrays';
 import { EventListener } from '../../../types';
 
 import { Event } from './event.class';