+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change Log
 All notable changes to the "Diff Folders" extension will be documented in this file.
 
+## [0.27.0] - 2020-07-19
+
+### Added
+- Added ignored files to list view.
+- Added new icon to toggle view for ignored files.
+- Added ignored files to stats.
+- Added `Select for Compare` and `Compare with Selected` for extension [Projects](https://marketplace.visualstudio.com/items?itemName=L13RARY.l13-projects)
+
+### Fixed
+- Fixed focus bug click into input fields.
+
+### Changed
+- Changed `${workspaceFolder:INDEX}` to `${workspaceFolder:NAME}`.
+
+### Removed
+- Removed `${workspaceFolder:INDEX}` and `${workspaceFolderBasename:NAME}`. Please use `${workspaceFolder:NAME}` instead.
+
 ## [0.26.0] - 2020-07-05
 
 ### Added
@@ -81,7 +98,7 @@ All notable changes to the "Diff Folders" extension will be documented in this f
 ## [0.21.0] - 2019-09-22
 
 ### Added
-- Added support for new extension [L13 Projects](https://marketplace.visualstudio.com/items?itemName=L13RARY.l13-projects)
+- Added support for new extension [Projects](https://marketplace.visualstudio.com/items?itemName=L13RARY.l13-projects)
 - Alt + Click on swap button changes the list view, too.
 
 ## [0.20.0] - 2019-08-25